+++ src/Resources/Users/index.ts
@@ -7,6 +7,7 @@ import Toybox from './Toybox';
 import Details from './Details';
 import Addresses from './Addresses';
 import ProductLists from './ProductLists';
+import Orders from './Orders';
 import { HttpClientConfig, HttpClientResponse } from '../../Interfaces/HttpClient';
 
 export default class Users extends AbstractResource {
@@ -17,6 +18,7 @@ export default class Users extends AbstractResource {
     public subscriptions: Subscriptions;
     public toybox: Toybox;
     public productLists: ProductLists;
+    public orders: Orders;
 
     constructor(api: Client) {
         super(api);
@@ -30,6 +32,7 @@ export default class Users extends AbstractResource {
         this.subscriptions = new Subscriptions(api);
         this.toybox = new Toybox(api);
         this.productLists = new ProductLists(api);
+        this.orders = new Orders(api);
     }
 
     initialise(): void {
@@ -46,5 +49,10 @@ export default class Users extends AbstractResource {
     public claimGuestResources: (...args: HttpClientConfig) => HttpClientResponse = this.createMethodFromPartialSpec({
         method: 'PATCH',
         path: '/claim',
+    });
+
+    public checkout: (...args: HttpClientConfig) => HttpClientResponse = this.createMethodFromPartialSpec({
+        method: 'POST',
+        path: '/checkout',
     });
 }